moved to graphql-tools for schema build
