+++ api/package.json
@@ -4,12 +4,13 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
-    "babel-preset-env": "^1.3.3",
+    "babel-preset-env": "^1.4.0",
     "babel-register": "^6.24.1",
     "bluebird": "^3.5.0",
     "express": "^4.15.2",
     "express-graphql": "^0.6.4",
-    "graphql": "^0.9.3"
+    "graphql": "^0.9.3",
+    "graphql-tools": "^0.11.0"
   },
   "devDependencies": {
     "babel-cli": "^6.24.1"
