+++ api/src/server.js
@@ -1,15 +1,18 @@
+import { makeExecutableSchema } from 'graphql-tools';
 import express from 'express';
 import graphqlHTTP from 'express-graphql';
-import schema from './schema';
-import rootValue from './rootValue';
+
+import typeDefs from './schema';
+import resolvers from './resolvers';
 
 const app = express();
+
+const schema = makeExecutableSchema({ typeDefs, resolvers });
 
 app.use(
   '/graphql',
   graphqlHTTP({
     schema,
-    rootValue,
     graphiql: true,
   }),
 );
