+++ package.json
@@ -4,5 +4,6 @@
   "main": "index.js",
   "author": "Stuart Harris <stuart.harris@red-badger.com>",
   "license": "MIT",
-  "devDependencies": {}
+  "devDependencies": {},
+  "dependencies": {}
 }
